wip value locations
