+++ wasm/jit/jit_value_location_test.go
@@ -1,6 +1,3 @@
-//go:build amd64
-// +build amd64
-
 package jit
 
 import (
